requirements.txt
cairocffi==1.0.2
CairoSVG==2.4.0
cffi==1.12.3
Click==7.0
cssselect2==0.2.1
defusedxml==0.6.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
Flask-WeasyPrint==0.6
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
pkg-resources==0.0.0
pycparser==2.19
pyodbc==4.0.26
Pyphen==0.9.5
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.6
tinycss2==1.0.2
WeasyPrint==48
webencodings==0.5.1
Werkzeug==0.15.5
